Splitting out server and app for additional coverage reporting
